Move properties for visibility from CefWebGui to OpenSpaceEngine
